Separated analysis code for easier use.
This makes running the analysis also outside AliPhysics easier.
